requirements.txt
Django==4.0.1
djoser==2.1.0
Pillow==9.0.0
asgiref==3.4.1
pywhatkit==5.3
gunicorn==20.1.0
whitenoise==5.3.0
virtualenv==20.13.0
python-decouple==3.5
python-dotenv==0.19.2
dj-database-url==0.5.0
psycopg2-binary==2.9.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
